bug fix: change type of courseNum in courseReg,and database bug fix: initialize numOfCurrStudent

+++ spl-net/src/main/java/bgu/spl/net/api/Messages/CourseRegMessage.java
@@ -4,9 +4,9 @@ import bgu.spl.net.api.Message;
 import bgu.spl.net.impl.RegistrationSystem.*;
 
 public class CourseRegMessage implements Message {
-    int courseNum;
+    Short courseNum;
 
-    public CourseRegMessage(int courseNum) {
+    public CourseRegMessage(Short courseNum) {
         this.courseNum = courseNum;
     }
